adding metric for total login attempts
issue: #92

this change adds a metric for the total number of login attempts.
it might seem that it's possible to sum both `keycloak_logins` and
`keycloak_failed_login_attempts` but as the `totalLogins` and
`totalFailedLoginAttempts` are counted separately their is now
guarantee that both counters contain the same metrics (labels with
same values). in order to make it possible to calculate the failure
rate per instance a new counter `totalLoginAttempts` is introduced.

Signed-off-by: xinau <felix.ehrenpfort@protonmail.com>