Fix Gradle build for v4.0.0 and Keycloak 22
The 4.0.0 release contains an update to Keycloak 22.
While this adjusted the dependencies in the pom.xml for the Maven
build, it did not adjust the build.gradle file.
This lead to './gradlew jar' failing.

We have now adjusted the build.gradle to reflect the same dependency
changes that have been made to the pom.xml.
With that, we also bump the Keycloak version in the gradle.properties
to 22.